% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendly2postGWAShelp.R
\name{chat_with_me}
\alias{chat_with_me}
\title{friendly2postGWAS包作者的联系方式}
\usage{
chat_with_me()
}
\value{
输入文件的GWAS数据格式和表头
}
\description{
friendly2postGWAS包作者的联系方式
}
\examples{
   chat_with_me()
}
