% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendly2postGWAShelp.R
\name{install_postGWAS_dependence}
\alias{install_postGWAS_dependence}
\title{安装friendly2postGWAS的相关依赖}
\usage{
install_postGWAS_dependence()
}
\value{
安装friendly2postGWAS的相关依赖包
}
\description{
安装friendly2postGWAS的相关依赖
}
\examples{
   install_postGWAS_dependence()
}
