% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendly2postGWAShelp.R
\name{install_postGWAS_R}
\alias{install_postGWAS_R}
\title{安装friendly2postGWAS包}
\usage{
install_postGWAS_R()
}
\value{
安装好friendly2postGWAS包
}
\description{
安装friendly2postGWAS包
}
\examples{
   install_postGWAS_R()
}
