% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendly2postGWAShelp.R
\name{postGWAS_update_plans}
\alias{postGWAS_update_plans}
\title{friendly2postGWAS包的后续更新计划}
\usage{
postGWAS_update_plans()
}
\value{
列出后续的friendly2postGWAS更新计划
}
\description{
friendly2postGWAS包的后续更新计划
}
\examples{
   postGWAS_update_plans()
}
